requirements.txt
ultralytics
opencv-python
streamlit
Pillow
ollama
fpdf2
numpy
plotly
pandas
